requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.32.5
SQLAlchemy==2.0.30